nikita/simple_tg_bot/requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
openai>=1.0.0 